chore: add Factory and Strategy patterns and break tokenApr calculation into files to better reading
